requirements/base.txt
Django==1.9.4
Unipath==1.1
psycopg2==2.6.1
django-braces==1.8.1
whitenoise==3.0
django-filter==0.13.0
django-mptt==0.8.3
django-rest-multiple-models==1.6.3
djangorestframework==3.3.3
Markdown==2.6.6
